+++ python/fedml/core/security/constants.py
@@ -9,8 +9,8 @@ DEFENSE_DP = "dp"
 DEFENSE_RFA = "rfa"
 DEFENSE_FOOLSGOLD = "foolsgold"
 DEFENSE_CRFL = "crfl"
-DEFENSE_THREESIGMA = "3sigma"
-DEFENSE_THREESIGMA_KRUM = "3sigma_krum"
+DEFENSE_THREESIGMA_FOOLSGOLD = "3sigma"
+DEFENSE_THREESIGMA = "3sigma_krum"
 DEFENSE_THREESIGMA_GEOMEDIAN = "3sigma_geo"
 DEFENSE_MULTIKRUM = "multikrum"
 DEFENSE_TRIMMED_MEAN = "trimmed_mean"
