+++ python/fedml/core/security/fedml_defender.py
@@ -6,9 +6,9 @@ from .defense.coordinate_wise_median_defense import CoordinateWiseMedianDefense
 from .defense.coordinate_wise_trimmed_mean_defense import CoordinateWiseTrimmedMeanDefense
 from .defense.crfl_defense import CRFLDefense
 from .defense.outlier_detection import OutlierDetection
-from .defense.three_sigma_defense import ThreeSigmaDefense
+from .defense.three_sigma_defense_foolsgold import ThreeSigmaDefense_Foolsgold
 from .defense.three_sigma_geomedian_defense import ThreeSigmaGeoMedianDefense
-from .defense.three_sigma_krum_defense import ThreeSigmaKrumDefense
+from .defense.three_sigma_defense import ThreeSigmaDefense
 from ..common.ml_engine_backend import MLEngineBackend
 from .defense.cclip_defense import CClipDefense
 from .defense.foolsgold_defense import FoolsGoldDefense
@@ -28,12 +28,12 @@ from ...core.security.constants import (
     DEFENSE_WEAK_DP,
     DEFENSE_RFA,
     DEFENSE_FOOLSGOLD,
-    DEFENSE_THREESIGMA,
+    DEFENSE_THREESIGMA_FOOLSGOLD,
     DEFENSE_CRFL,
     DEFENSE_MULTIKRUM,
     DEFENSE_TRIMMED_MEAN,
     DEFENSE_THREESIGMA_GEOMEDIAN,
-    DEFENSE_THREESIGMA_KRUM, ANOMALY_DETECTION, DEFENSE_WISE_MEDIAN, DEFENSE_DIFF_CLIPPING,
+    DEFENSE_THREESIGMA, ANOMALY_DETECTION, DEFENSE_WISE_MEDIAN, DEFENSE_DIFF_CLIPPING,
 )
 
 
@@ -80,12 +80,12 @@ class FedMLDefender:
                 self.defender = RFADefense(args)
             elif self.defense_type == DEFENSE_FOOLSGOLD:
                 self.defender = FoolsGoldDefense(args)
-            elif self.defense_type == DEFENSE_THREESIGMA:
-                self.defender = ThreeSigmaDefense(args)
+            elif self.defense_type == DEFENSE_THREESIGMA_FOOLSGOLD:
+                self.defender = ThreeSigmaDefense_Foolsgold(args)
             elif self.defense_type == DEFENSE_THREESIGMA_GEOMEDIAN:
                 self.defender = ThreeSigmaGeoMedianDefense(args)
-            elif self.defense_type == DEFENSE_THREESIGMA_KRUM:
-                self.defender = ThreeSigmaKrumDefense(args)
+            elif self.defense_type == DEFENSE_THREESIGMA:
+                self.defender = ThreeSigmaDefense(args)
             elif self.defense_type == DEFENSE_CRFL:
                 self.defender = CRFLDefense(args)
             elif self.defense_type == DEFENSE_TRIMMED_MEAN:
@@ -135,9 +135,9 @@ class FedMLDefender:
         return self.is_defense_enabled() and self.defense_type in [
             DEFENSE_SLSGD,
             DEFENSE_FOOLSGOLD,
-            DEFENSE_THREESIGMA,
+            DEFENSE_THREESIGMA_FOOLSGOLD,
             DEFENSE_THREESIGMA_GEOMEDIAN,
-            DEFENSE_THREESIGMA_KRUM,
+            DEFENSE_THREESIGMA,
             DEFENSE_KRUM,
             DEFENSE_CCLIP,
             DEFENSE_MULTIKRUM,
